Use reflected implementation of SendMessage method
