+++ ReportPortal.NUnitExtension/ReportPortal.NUnitExtension.csproj
@@ -16,7 +16,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="NUnit" Version="3.11.0" />
     <PackageReference Include="NUnit.Engine.Api" Version="3.9.0">
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
